refactor(web/routes): extract conversations and API route modules from server.ts
Split the 1695-line src/web/server.ts into focused modules:

- src/web/middleware/auth.ts: shared auth helpers (SESSION_COOKIE,
  sessionAuthMiddleware, optionalAuthMiddleware, adminGuard,
  getUserFilterIds, isConversationOwner, buildCookieOptions)
- src/web/routes/helpers.ts: pure request helpers (parsePagination,
  attachTags)
- src/web/routes/conversations.ts: all /c/* route handlers
- src/web/routes/api.ts: all /api/* + /notifications handlers plus
  event-bus SSE wiring

server.ts now mounts these routers and shrinks from 1695 to 720 lines.
No route URLs, response shapes, or exported function signatures changed.
